+++ sora-android-sdk/src/main/kotlin/jp/shiguredo/sora/sdk/channel/rtc/RTCLocalAudioManager.kt
@@ -1,32 +1,26 @@
 package jp.shiguredo.sora.sdk.channel.rtc
 
+import jp.shiguredo.sora.sdk.channel.option.SoraAudioOption
 import jp.shiguredo.sora.sdk.util.SoraLogger
 import org.webrtc.*
 import java.util.*
 
+
 class RTCLocalAudioManager(
-        private val send:         Boolean,
-        private val levelControl: Boolean = false,
-        private val processing:   Boolean = false
+        private val send:         Boolean
 ) {
 
     companion object {
         private val TAG = RTCLocalAudioManager::class.simpleName
-
-        private const val ECHO_CANCELLATION_CONSTRAINT = "googEchoCancellation"
-        private const val AUTO_GAIN_CONTROL_CONSTRAINT = "googAutoGainControl"
-        private const val HIGH_PASS_FILTER_CONSTRAINT  = "googHighpassFilter"
-        private const val NOISE_SUPPRESSION_CONSTRAINT = "googNoiseSuppression"
-        private const val LEVEL_CONTROL_CONSTRAINT     = "levelControl"
     }
 
     private var source: AudioSource? = null
     private var track:  AudioTrack?  = null
 
-    fun initTrack(factory: PeerConnectionFactory) {
+    fun initTrack(factory: PeerConnectionFactory, audioOption: SoraAudioOption) {
         SoraLogger.d(TAG, "initTrack")
         if (send) {
-            val constraints = createSourceConstraints()
+            val constraints = createSourceConstraints(audioOption)
             source = factory.createAudioSource(constraints)
             SoraLogger.d(TAG, "audio source created: ${source}")
             val trackId = UUID.randomUUID().toString()
@@ -36,21 +30,23 @@ class RTCLocalAudioManager(
         }
     }
 
-    private fun createSourceConstraints(): MediaConstraints {
+    private fun createSourceConstraints(audioOption: SoraAudioOption): MediaConstraints {
         val constraints = MediaConstraints()
-        if (levelControl) {
+        if (!audioOption.audioProcessingEchoCancellation) {
             constraints.mandatory.add(
-                    MediaConstraints.KeyValuePair(LEVEL_CONTROL_CONSTRAINT, "true"))
+                    MediaConstraints.KeyValuePair(SoraAudioOption.ECHO_CANCELLATION_CONSTRAINT, "false"))
         }
-        if (!processing) {
+        if(!audioOption.audioProcessingAutoGainControl) {
             constraints.mandatory.add(
-                    MediaConstraints.KeyValuePair(ECHO_CANCELLATION_CONSTRAINT, "false"))
+                    MediaConstraints.KeyValuePair(SoraAudioOption.AUTO_GAIN_CONTROL_CONSTRAINT, "false"))
+        }
+        if (!audioOption.audioProcessingHighpassFilter) {
             constraints.mandatory.add(
-                    MediaConstraints.KeyValuePair(AUTO_GAIN_CONTROL_CONSTRAINT, "false"))
+                    MediaConstraints.KeyValuePair(SoraAudioOption.HIGH_PASS_FILTER_CONSTRAINT, "false"))
+        }
+        if (!audioOption.audioProcessingNoiseSuppression) {
             constraints.mandatory.add(
-                    MediaConstraints.KeyValuePair(HIGH_PASS_FILTER_CONSTRAINT, "false"))
-            constraints.mandatory.add(
-                    MediaConstraints.KeyValuePair(NOISE_SUPPRESSION_CONSTRAINT, "false"))
+                    MediaConstraints.KeyValuePair(SoraAudioOption.NOISE_SUPPRESSION_CONSTRAINT, "false"))
         }
         return constraints
     }