+++ sora-android-sdk/src/main/kotlin/jp/shiguredo/sora/sdk/channel/signaling/SignalingChannel.kt
@@ -2,10 +2,7 @@ package jp.shiguredo.sora.sdk.channel.signaling
 
 import jp.shiguredo.sora.sdk.channel.option.SoraChannelRole
 import jp.shiguredo.sora.sdk.channel.option.SoraMediaOption
-import jp.shiguredo.sora.sdk.channel.signaling.message.MessageConverter
-import jp.shiguredo.sora.sdk.channel.signaling.message.NotificationMessage
-import jp.shiguredo.sora.sdk.channel.signaling.message.OfferConfig
-import jp.shiguredo.sora.sdk.channel.signaling.message.PushMessage
+import jp.shiguredo.sora.sdk.channel.signaling.message.*
 import jp.shiguredo.sora.sdk.error.SoraErrorReason
 import jp.shiguredo.sora.sdk.util.SoraLogger
 import okhttp3.*
@@ -25,7 +22,7 @@ interface SignalingChannel {
     interface Listener {
         fun onConnect()
         fun onDisconnect()
-        fun onInitialOffer(clientId: String, sdp: String, config: OfferConfig?)
+        fun onInitialOffer(offerMessage: OfferMessage)
         fun onUpdatedOffer(sdp: String)
         fun onReOffer(sdp: String)
         fun onError(reason: SoraErrorReason)
@@ -70,8 +67,6 @@ class SignalingChannelImpl @JvmOverloads constructor(
             SoraLogger.i(TAG, "signaling is closing")
             return
         }
-
-        SoraLogger.d(TAG, sdp)
 
         webSocket?.let {
             val msg = MessageConverter.buildAnswerMessage(sdp)
@@ -154,7 +149,6 @@ class SignalingChannelImpl @JvmOverloads constructor(
                     clientId                = clientId,
                     signalingNotifyMetadata = signalingNotifyMetadata
             )
-            SoraLogger.d(TAG, "connect message: $message")
             it.send(message)
         }
     }
@@ -165,21 +159,12 @@ class SignalingChannelImpl @JvmOverloads constructor(
     }
 
     private fun onOfferMessage(text: String) {
-        val offer = MessageConverter.parseOfferMessage(text)
+        val offerMessage = MessageConverter.parseOfferMessage(text)
+        SoraLogger.d(TAG, """[signaling:$role] <- offer
+            |${offerMessage.sdp}""".trimMargin())
         // TODO message validation
 
-        SoraLogger.d(TAG, "[signaling:$role] <- offer")
-        SoraLogger.d(TAG, offer.sdp)
-
-        // connectionId は Sora 19.04.0 から返ってくるため、ない場合は clientId を使う
-        val connectionId = when {
-            offer.connectionId != null ->
-                offer.connectionId
-            else ->
-                offer.clientId
-        }
-
-        listener?.onInitialOffer(connectionId, offer.sdp, offer.config)
+        listener?.onInitialOffer(offerMessage)
     }
 
     private fun onUpdateMessage(text: String) {